+++ src/pyscaffold/utils.py
@@ -17,10 +17,7 @@ from operator import itemgetter
 from .exceptions import InvalidIdentifier, OldSetuptools
 from .log import logger
 from .templates import licenses
-
-# Taken from six to avoid external import
-PY2 = sys.version_info[0] == 2
-PY3 = sys.version_info[0] == 3
+from .contrib.six import PY2
 
 
 @contextmanager